requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.9
django-cors-headers==3.11.0
django-filter==21.1
django-password-reset==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Markdown==3.3.4
Pillow==8.4.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
